updateProjektarray auskomm: Löscht Projektordner
Leseboxvorbereiten: Lesebox checken. Projekte, die in der PList fehlen, nachtragen.
> !!!! Machte Mühe mit Stringvergleich bei Umlauten

init self ArchivView: Verwechslung von View und Daten korr

Umlaut-Texte mit stringWithUTF8String ergänzt ???

Utils: Ordnernamenanpfad erstellt eine Liste der Unterordner an einem Pfad

+++ RecPlayAdminPlayer/rExportKontroller.m
@@ -58,7 +58,7 @@ NSString*	RPExportformatKey;
  return[RPExportdaten length];
  }
  
- - (int)ExportPrefsSchreiben
+ - (long)ExportPrefsSchreiben
  {
  	short l=[RPExportdaten length];
 	if(l>0)
@@ -532,7 +532,7 @@ NSString*	RPExportformatKey;
 }
 
 
-- (int) AufnahmeExportierenMitPfad:(NSString*)derAufnahmePfad 
+- (long) AufnahmeExportierenMitPfad:(NSString*)derAufnahmePfad
 					 mitUserDialog:(BOOL)userDialogOK
 				 mitSettingsDialog:(BOOL)settingsDialogOK
 {
